+++ app/models/grandprix_event.rb
@@ -3,4 +3,8 @@ class GrandprixEvent < ApplicationRecord
   has_many :skaters, through: :grandprix_entries
 
   belongs_to :category
+
+  scope :done, -> { where(done: true) }
+  scope :incoming, -> { where(done: false) }
+    
 end